Fix move ctor/assign.
The parse stack usually contains pointers into the singletons_
object. After move-construction or move-assignment, we now adjust
those pointers so that they point to the new instance rather than the
original.

The 'singletons' type no longer uses aligned_storage<> but directly
instantiates the eof, trailing_ws, and root instances that will
certainly be needed as the initial stack is built.