Use get_if<> rather than std::get<> for the large type.
We go to some lengths to ensure that the variant cannot end up in the
"valueless" state by ensuring that the ctors for both types that it
can hold are noexcept. That means that one of get_if<small_type> or
get_if<large_type> must always succeed. This allows us to bypass the
check (and potential exception throw) that is performed by
std::get<>.