Add "enum". Clean up schema.cpp. Floats can sometimes be cast to int
Floating point values with a 0 fractional component and within the
range of the integer types can be cast to integer. This allows the
code to treat "1.0" and "1" as being the same. Previously, the first
of these was a float and the second and integer.

Remove the test code from schema.cpp and report errors from
schema::check() to the user.