Workaround for Cov CID1571023. (#41)
Coverity thinks that there is a path which does not check that the
call to find() will not return end(). In fact, the previous ASSERT
should already have checked this, but however... This adds an
additional use of ASSERT_NE to guarantee the validity of the iterator
returned by find().