small_vector CTAD / eliminate transfer_from_same() (#50)
* Add a simple CTAD guide to small_vector<>.

* Eliminate the transfer_from_same() members.

Coverity CID1569248 is under the impression that the
transfer_from_same() function could throw if the variant was storing a
type with index other than the function's Index template
argument. The two being the same was, however, a precondition of the
function.

Since the two versions of transfer_from_same() was called only once
each, I think that inlining them in the location that the precondition
is established makes more sense than having them as separate functions.