Clean up building of the clang warning suppression.
Some of the warning suppression flags are not available in all
versions of clang. This means that we need to test the compiler's
support for each before adding ths switch.

This change achieves this useing a list of flags and a function rather
than the ad-hoc list of checks used previously.